# Auxiliary qubit tracking in ``HighLevelSynthesis``