--- a/qiskit/transpiler/preset_passmanagers/generate_preset_pass_manager.py
+++ b/qiskit/transpiler/preset_passmanagers/generate_preset_pass_manager.py
@@ -59,6 +59,7 @@ def generate_preset_pass_manager(
     init_method=None,
     optimization_method=None,
     dt=None,
+    qubits_initially_zero=True,
     *,
     _skip_target=False,
 ):
@@ -233,6 +234,8 @@ def generate_preset_pass_manager(
             plugin is not used. You can see a list of installed plugins by
             using :func:`~.list_stage_plugins` with ``"optimization"`` for the
             ``stage_name`` argument.
+        qubits_initially_zero (bool): Indicates whether the input circuit is
+                zero-initialized.
 
     Returns:
         StagedPassManager: The preset pass manager for the given options
@@ -376,6 +379,7 @@ def generate_preset_pass_manager(
         "hls_config": hls_config,
         "init_method": init_method,
         "optimization_method": optimization_method,
+        "qubits_initially_zero": qubits_initially_zero,
     }
 
     if backend is not None: