… qiskit/transpiler/preset_passmanagers/builtin_plugins.py
… qiskit/transpiler/preset_passmanagers/common.py
… releasenotes/notes/hls-with-ancillas-d6792b41dfcf4aac.yaml
… test/python/transpiler/test_high_level_synthesis.py
… test/python/transpiler/test_passmanager_config.py