<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add a release note file using reno if the change needs to be
  documented in the release notes.
⚠️ If your pull request fixes an open issue, please link to the issue.

- [ ] I have added the tests to cover my changes.
- [ ] I have updated the documentation accordingly.
- [ ] I have read the CONTRIBUTING document.
-->

### Summary

Enable tracking the state of auxiliary qubits in high-level synthesis. This allows synthesis plugins to leverage additional qubits in a decomposition, leading to a potential reduction of circuit depth and number of operations (which is why I tagged it for the performance demo 🙂). This PR unblocks work on the circuit library, in particular refactoring of the MCX and MCMT gates.

### Details and comments

This first implementation follows a greedy approach, where each operation is given access to all currently available auxiliary qubits. It does so by using the newly added `QubitTracker`, which keeps track of which qubits in the circuit/dag are "clean" (in state $|0\rangle$) or "dirty" (an unknown state). Generally, if qubits start out as clean, and as soon as it is involved in an operation, it changes to dirty (special rules apply for operations like resets, barriers or delays).

Another important change is that synthesized blocks must be synthesized again. This recursion is necessary, since unrolling or decomposing an operation, could yield other operations that must be synthesized. This slows down the code compared to the current state, but only by another iteration over the DAG, so that should be a reasonably small overhead. 

Follow-ups could involve:
* speedup the code by having the synthesis plugins return DAG circuits (will be easier once we could construct a DAG from Rust's `CircuitData` or an iterator over circuit instructions)
* avoid redundant recursion if plugins can report the basis gate set of their decomposition
* do the whole thing in Rust 🙂 

<details closed>
<summary>The asv utility benchmark finds that this does not significantly reduce performance.</summary>
<br>

```
Benchmarks that have stayed the same:

| Change   | Before [1fdd527f] <main>   | After [9726d19b] <hls+aux~1>   | Ratio   | Benchmark (Parameter)                                                         |
|----------|----------------------------|--------------------------------|---------|-------------------------------------------------------------------------------|
|          | 0                          | 0                              | n/a     | utility_scale.UtilityScaleBenchmarks.track_bvlike_depth('cx')                 |
|          | 0                          | 0                              | n/a     | utility_scale.UtilityScaleBenchmarks.track_bvlike_depth('cz')                 |
|          | 0                          | 0                              | n/a     | utility_scale.UtilityScaleBenchmarks.track_bvlike_depth('ecr')                |
|          | 701±6ms                    | 756±6ms                        | 1.08    | utility_scale.UtilityScaleBenchmarks.time_square_heisenberg('cx')             |
|          | 877±9ms                    | 939±10ms                       | 1.07    | utility_scale.UtilityScaleBenchmarks.time_square_heisenberg('ecr')            |
|          | 3.40±0.02s                 | 3.61±0.08s                     | 1.06    | utility_scale.UtilityScaleBenchmarks.time_qv('ecr')                           |
|          | 3.33±0.01s                 | 3.51±0.08s                     | 1.05    | utility_scale.UtilityScaleBenchmarks.time_qv('cx')                            |
|          | 5.46±0s                    | 5.74±0.06s                     | 1.05    | utility_scale.UtilityScaleBenchmarks.time_qv('cz')                            |
|          | 1.02±0.02s                 | 1.07±0.01s                     | 1.05    | utility_scale.UtilityScaleBenchmarks.time_square_heisenberg('cz')             |
|          | 223±5ms                    | 231±9ms                        | 1.04    | utility_scale.UtilityScaleBenchmarks.time_bv_100('cx')                        |
|          | 1.38±0.01s                 | 1.43±0.01s                     | 1.04    | utility_scale.UtilityScaleBenchmarks.time_qaoa('ecr')                         |
|          | 248±2ms                    | 256±3ms                        | 1.03    | utility_scale.UtilityScaleBenchmarks.time_bv_100('cz')                        |
|          | 15.3±0.01s                 | 15.7±0.1s                      | 1.03    | utility_scale.UtilityScaleBenchmarks.time_qft('ecr')                          |
|          | 238±1ms                    | 243±2ms                        | 1.02    | utility_scale.UtilityScaleBenchmarks.time_bv_100('ecr')                       |
|          | 6.81±0.03ms                | 6.93±0.2ms                     | 1.02    | utility_scale.UtilityScaleBenchmarks.time_parse_qaoa_n100('cx')               |
|          | 917±5ms                    | 931±6ms                        | 1.02    | utility_scale.UtilityScaleBenchmarks.time_qaoa('cx')                          |
|          | 73.4±0.4ms                 | 74.0±0.2ms                     | 1.01    | utility_scale.UtilityScaleBenchmarks.time_bvlike('cx')                        |
|          | 73.5±0.2ms                 | 74.1±0.6ms                     | 1.01    | utility_scale.UtilityScaleBenchmarks.time_bvlike('cz')                        |
|          | 73.3±0.2ms                 | 73.7±0.3ms                     | 1.01    | utility_scale.UtilityScaleBenchmarks.time_bvlike('ecr')                       |
|          | 6.78±0.05ms                | 6.86±0.08ms                    | 1.01    | utility_scale.UtilityScaleBenchmarks.time_parse_qaoa_n100('cz')               |
|          | 6.86±0.08ms                | 6.90±0.05ms                    | 1.01    | utility_scale.UtilityScaleBenchmarks.time_parse_qaoa_n100('ecr')              |
|          | 73.0±0.4ms                 | 73.5±0.3ms                     | 1.01    | utility_scale.UtilityScaleBenchmarks.time_parse_qft_n100('cx')                |
|          | 23.3±0.2ms                 | 23.5±0.2ms                     | 1.01    | utility_scale.UtilityScaleBenchmarks.time_parse_square_heisenberg_n100('cx')  |
|          | 23.3±0.09ms                | 23.7±0.3ms                     | 1.01    | utility_scale.UtilityScaleBenchmarks.time_parse_square_heisenberg_n100('cz')  |
|          | 1.78±0.01s                 | 1.80±0.01s                     | 1.01    | utility_scale.UtilityScaleBenchmarks.time_qaoa('cz')                          |
|          | 14.7±0.03s                 | 14.8±0.01s                     | 1.01    | utility_scale.UtilityScaleBenchmarks.time_qft('cx')                           |
|          | 15.8±0.01s                 | 16.1±0.03s                     | 1.01    | utility_scale.UtilityScaleBenchmarks.time_qft('cz')                           |
|          | 3.84±0.01s                 | 3.84±0.01s                     | 1.00    | utility_scale.UtilityScaleBenchmarks.time_circSU2('cz')                       |
|          | 395                        | 395                            | 1.00    | utility_scale.UtilityScaleBenchmarks.track_bv_100_depth('cx')                 |
|          | 397                        | 397                            | 1.00    | utility_scale.UtilityScaleBenchmarks.track_bv_100_depth('cz')                 |
|          | 397                        | 397                            | 1.00    | utility_scale.UtilityScaleBenchmarks.track_bv_100_depth('ecr')                |
|          | 300                        | 300                            | 1.00    | utility_scale.UtilityScaleBenchmarks.track_circSU2_depth('cx')                |
|          | 300                        | 300                            | 1.00    | utility_scale.UtilityScaleBenchmarks.track_circSU2_depth('cz')                |
|          | 300                        | 300                            | 1.00    | utility_scale.UtilityScaleBenchmarks.track_circSU2_depth('ecr')               |
|          | 1483                       | 1483                           | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qaoa_depth('cx')                   |
|          | 1488                       | 1488                           | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qaoa_depth('cz')                   |
|          | 1488                       | 1488                           | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qaoa_depth('ecr')                  |
|          | 1954                       | 1954                           | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qft_depth('cx')                    |
|          | 1954                       | 1954                           | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qft_depth('cz')                    |
|          | 1954                       | 1954                           | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qft_depth('ecr')                   |
|          | 2538                       | 2538                           | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qv_depth('cx')                     |
|          | 2538                       | 2538                           | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qv_depth('cz')                     |
|          | 2538                       | 2538                           | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qv_depth('ecr')                    |
|          | 435                        | 435                            | 1.00    | utility_scale.UtilityScaleBenchmarks.track_square_heisenberg_depth('cx')      |
|          | 435                        | 435                            | 1.00    | utility_scale.UtilityScaleBenchmarks.track_square_heisenberg_depth('cz')      |
|          | 435                        | 435                            | 1.00    | utility_scale.UtilityScaleBenchmarks.track_square_heisenberg_depth('ecr')     |
|          | 3.85±0.06s                 | 3.80±0.05s                     | 0.99    | utility_scale.UtilityScaleBenchmarks.time_circSU2('ecr')                      |
|          | 73.1±0.9ms                 | 72.7±0.8ms                     | 0.99    | utility_scale.UtilityScaleBenchmarks.time_parse_qft_n100('cz')                |
|          | 73.0±0.5ms                 | 72.4±0.6ms                     | 0.99    | utility_scale.UtilityScaleBenchmarks.time_parse_qft_n100('ecr')               |
|          | 23.4±0.3ms                 | 23.2±0.1ms                     | 0.99    | utility_scale.UtilityScaleBenchmarks.time_parse_square_heisenberg_n100('ecr') |
|          | 3.78±0.1s                  | 3.67±0.02s                     | 0.97    | utility_scale.UtilityScaleBenchmarks.time_circSU2('cx')                       |

BENCHMARKS NOT SIGNIFICANTLY CHANGED.
```
<\details>